+++ src/Formie.php
@@ -115,7 +115,7 @@ class Formie extends Plugin
 
     public bool $hasCpSection = true;
     public bool $hasCpSettings = true;
-    public string $schemaVersion = '4.0.47';
+    public string $schemaVersion = '4.0.48';
     public string $minVersionRequired = '2.1.5';